# swagger.yaml
basePath: /
consumes:
- application/json
definitions:
  Branch:
    description: Branch represents a delivery location
    properties:
      address:
        type: string
        x-go-name: Address
      branchID:
        format: int64
        type: integer
        x-go-name: BranchID
      latitude:
        format: float
        type: number
        x-go-name: Latitude
      longitude:
        format: float
        type: number
        x-go-name: Longitude
    type: object
    x-go-package: github.com/miguelapabenedit/fravega-challange/branch
host: localhost:5000/api/v1
info:
  description: Documentation of our awesome API.
  title: awesome.
  version: 1.0.0
paths:
  /branch:
    get:
      description: Gets the nearest branch for deliver base on a position
      operationId: currentPosition
      parameters:
      - description: required
        format: float
        in: query
        name: latitude
        type: number
        x-go-name: Latitude
      - description: required
        format: float
        in: query
        name: longitude
        type: number
        x-go-name: Longitude
      responses:
        "200":
          $ref: '#/responses/branchResponse'
        "400":
          $ref: '#/responses/http.BadRequest'
    post:
      description: Creates a new branch
      operationId: branchParams
      parameters:
      - description: required
        in: body
        name: Body
        schema:
          $ref: '#/definitions/Branch'
      responses:
        "200":
          description: ""
        "400":
          description: ""
  /branch/{branchID}:
    get:
      description: Gets a branch base on an branchId
      operationId: branchId
      parameters:
      - description: required
        format: int32
        in: path
        name: branchID
        required: true
        type: integer
        x-go-name: BranchID
      responses:
        "200":
          $ref: '#/responses/branchResponse'
        "400":
          $ref: '#/responses/http.BadRequest'
produces:
- application/json
responses:
  branchResponse:
    description: ""
    schema:
      $ref: '#/definitions/Branch'
schemes:
- http
swagger: "2.0"